+++ massiv/tests/Data/Massiv/Array/Ops/SliceSpec.hs
@@ -17,7 +17,7 @@ import           Test.QuickCheck
 -- extract
 
 prop_ExtractEqualsExtractFromTo
-  :: (Eq (Array (EltRepr r ix) ix e), Arbitrary (Array r ix e), Size r ix e)
+  :: (Eq (Array (EltRepr r ix) ix e), Arbitrary (Array r ix e), Extract r ix e)
   => proxy (r, ix, e) -> SzIx ix -> Array r ix e -> Bool
 prop_ExtractEqualsExtractFromTo _ (SzIx (Sz eIx) sIx) arr =
   extractFromTo sIx eIx arr == extract sIx (liftIndex2 (-) eIx sIx) arr
@@ -26,7 +26,7 @@ prop_ExtractEqualsExtractFromTo _ (SzIx (Sz eIx) sIx) arr =
 
 
 specSizeN
-  :: (Eq (Array (EltRepr r ix) ix e), Arbitrary (Array r ix e), Show (Array r ix e), Arbitrary ix, Size r ix e)
+  :: (Eq (Array (EltRepr r ix) ix e), Arbitrary (Array r ix e), Show (Array r ix e), Arbitrary ix, Extract r ix e)
   => proxy (r, ix, e) -> Spec
 specSizeN proxy = do
   describe "extract" $ do